Add endpoints for dataloaders to /meta
