+++ tests/api/routers/test_meta.py
@@ -1,0 +1,19 @@
+import pytest
+from toktagger.api.schemas.samples import ShotData
+
+
+@pytest.mark.asyncio
+async def test_get_data_loaders(api_client, setup_db):
+    response = await api_client.get("/meta/dataloader")
+    assert response.status_code == 200
+    data = response.json()
+    assert all(item in data for item in ("uda", "image", "parquet"))
+
+
+@pytest.mark.asyncio
+async def test_get_data_schema(api_client, setup_db):
+    response = await api_client.get("/meta/dataloader/uda")
+    assert response.status_code == 200
+    data = response.json()
+    assert data["title"] == "ShotData"
+    assert data == ShotData.model_json_schema()
